move server items related to client to 'client' directory
