+++ src/block_device_monitor.cpp
@@ -12,7 +12,7 @@
 #include "common/autos/auto_mount.hpp"
 #include "common/autos/auto_temp_folder.hpp"
 #include "common/os_utils.hpp"
-#include "rules_utils.hpp"
+#include "client/rules_utils.hpp"
 
 BlockDeviceMonitor::BlockDeviceMonitor(std::unique_ptr<IReader> event_reader, std::unique_ptr<ClientAccepter> client_accepter)
     : _event_reader(std::move(event_reader)), _client_accepter(std::move(client_accepter)), _clients_lock(std::make_shared<std::mutex>()), _rules_manager(std::make_shared<RulesManager>())
