+++ src/block_device_monitor.hpp
@@ -6,10 +6,10 @@ using namespace std::filesystem;
 #include "common/writers/writer.hpp"
 #include "common/readers/reader.hpp"
 #include "udev_event.hpp"
-#include "client.hpp"
-#include "client_handler.hpp"
-#include "client_accepter.hpp"
-#include "rules_manager.hpp"
+#include "client/client.hpp"
+#include "client/client_handler.hpp"
+#include "client/client_accepter.hpp"
+#include "client/rules_manager.hpp"
 #include "proto/rule.pb.h"
 #include "block_device.hpp"
