+++ src/client/client_accepter.hpp
@@ -1,5 +1,5 @@
-#include "common/network/server_socket.hpp"
 #include "client.hpp"
+#include "../common/network/server_socket.hpp"
 
 static constexpr size_t DEFAULT_MAX_CONCURRENT_CLIENTS = 10;
